image-organizer/backend/requirements.txt
google-generativeai
flask
flask-cors
pillow
python-dotenv
werkzeug